requirements.txt
ollama
streamlit
PyPDF2
sentence-transformers
faiss-cpu